+++ backend/src/entity/Notifier.ts
@@ -1,2 +1,10 @@
-// id 
-// id do ponto de interesse
+import { Column, Entity, PrimaryGeneratedColumn } from "typeorm";
+
+@Entity()
+export class Notifier {
+  @PrimaryGeneratedColumn()
+  id: number;
+
+  @Column()
+  benchmarkId: number;
+}
